feat: Give support to both Hailo8 and Hailo10
