+++ .gitignore
@@ -1,11 +1,13 @@
 # The directory Mix will write compiled artifacts to.
 /_build/
+_build/
 
 # If you run "mix test --cover", coverage assets end up here.
 /cover/
 
 # The directory Mix downloads your dependencies sources to.
 /deps/
+deps/
 
 # Where third-party dependencies like ExDoc output generated docs.
 /doc/
