+++ priv/.hailo_target
@@ -1,0 +1,1 @@
+hailo10
